Show chars instead of numbers. Set font and size. Chars are still not loaded from file but in a static array. Enlarge view. Fix warning with casting

+++ DMLazyScrollView/DMLazyScrollView.m
@@ -148,7 +148,7 @@ enum {
 - (void) setNumberOfPages:(NSUInteger)pages {
     if (pages != numberOfPages) {
         numberOfPages = pages;
-        int offset = [self hasMultiplePages] ? numberOfPages + 2 : 1;
+        int offset = [self hasMultiplePages] ? (int)numberOfPages + 2 : 1;
         if (_direction == DMLazyScrollViewDirectionHorizontal) {
             self.contentSize = CGSizeMake(self.frame.size.width * offset,
                                           self.contentSize.height);
